move and update Item component
