+++ web/src/beta/components/ListItem/index.stories.tsx
@@ -1,0 +1,15 @@
+import { Meta, StoryObj } from "@storybook/react";
+
+import ActionItem from "./index";
+
+export default {
+  component: ActionItem,
+} as Meta;
+
+type Story = StoryObj<typeof ActionItem>;
+
+export const Default: Story = {
+  args: {
+    text: "long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text ",
+  },
+};
