--- web/src/beta/features/Editor/tabs/story/SidePanel/Item/index.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from "@storybook/react";
-
-import ActionItem from "./index";
-
-export default {
-  component: ActionItem,
-} as Meta;
-
-type Story = StoryObj<typeof ActionItem>;
-
-export const Default: Story = {
-  args: {
-    children:
-      "long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text long text ",
-  },
-};
